backend/requirements.txt
flask
python-dotenv
sqlalchemy
werkzeug